requirements.txt
Django==2.1.12
djangorestframework==3.10.3
django-redis==4.10.0
psycopg2-binary==2.8.6
django-cors-headers==3.3.0
PyJWT==1.7.1
uWSGI==2.0.18
celery==4.3.0